move firstboot from snappy/ to firstboot/
