+++ firstboot/firstboot.go
@@ -20,10 +20,80 @@
 package firstboot
 
 import (
+	"errors"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/osutil"
 )
 
 func HasRun() bool {
 	return osutil.FileExists(dirs.SnapFirstBootStamp)
+}
+
+var (
+	// ErrNotFirstBoot is an error that indicates that the first boot has already
+	// run
+	ErrNotFirstBoot = errors.New("this is not your first boot")
+)
+
+// FirstBoot checks whether it's the first boot, and if so enables the
+// first ethernet device and runs gadgetConfig (as well as flagging that
+// it run)
+func FirstBoot() error {
+	if HasRun() {
+		return ErrNotFirstBoot
+	}
+	defer stampFirstBoot()
+
+	return enableFirstEther()
+}
+
+func stampFirstBoot() error {
+	// filepath.Dir instead of firstbootDir directly to ease testing
+	stampDir := filepath.Dir(dirs.SnapFirstBootStamp)
+
+	if _, err := os.Stat(stampDir); os.IsNotExist(err) {
+		if err := os.MkdirAll(stampDir, 0755); err != nil {
+			return err
+		}
+	}
+
+	return osutil.AtomicWriteFile(dirs.SnapFirstBootStamp, []byte{}, 0644, 0)
+}
+
+var globs = []string{"/sys/class/net/eth*", "/sys/class/net/en*"}
+var ethdir = "/etc/network/interfaces.d"
+var ifup = "/sbin/ifup"
+
+func enableFirstEther() error {
+	var eths []string
+	for _, glob := range globs {
+		eths, _ = filepath.Glob(glob)
+		if len(eths) != 0 {
+			break
+		}
+	}
+	if len(eths) == 0 {
+		return nil
+	}
+	eth := filepath.Base(eths[0])
+	ethfile := filepath.Join(ethdir, eth)
+	data := fmt.Sprintf("allow-hotplug %[1]s\niface %[1]s inet dhcp\n", eth)
+
+	if err := osutil.AtomicWriteFile(ethfile, []byte(data), 0644, 0); err != nil {
+		return err
+	}
+
+	ifup := exec.Command(ifup, eth)
+	ifup.Stdout = os.Stdout
+	ifup.Stderr = os.Stderr
+	if err := ifup.Run(); err != nil {
+		return err
+	}
+
+	return nil
 }
